more refactoring for evaluation procedure + add first try for poincare

+++ utils/utils.py
@@ -42,21 +42,24 @@ def load_network_from_checkpoint(network, epoch, name_prefix_for_saved_model, st
 
 
 def get_labels_matrix(labels_list_1, labels_list_2):
-    matrix = torch.from_numpy(np.zeros((len(labels_list_1), len(labels_list_2))))
+    matrix = torch.from_numpy(np.zeros((len(labels_list_1), len(labels_list_2)), dtype=int))
     for i, labels1 in enumerate(labels_list_1):
         for j, labels2 in enumerate(labels_list_2):
             # print('labels1 ', labels1)
             # print('labels2 ', labels2)
+
+            # removing dummy labels which were added in order to have exactly 3 labels for each image
             new_labels_1 = np.delete(labels1.numpy(), np.where(labels1.numpy() == 0)[0])
             new_labels_2 = np.delete(labels2.numpy(), np.where(labels2.numpy() == 0)[0])
-            # print('new_labels1 ', new_labels_1)
-            # print('new_labels2 ', new_labels_2)
+            #print('new_labels1 ', new_labels_1, 'new_labels2 ', new_labels_2)
             # print('np.in1d(labels1, labels2) ', np.in1d(new_labels_1, new_labels_2))
             # print('np.in1d(labels1, labels2).any() ', np.in1d(new_labels_1, new_labels_2).any())
             # print('int(np.in1d(labels1.numpy(), labels2.numpy()).any()) ', int(np.in1d(new_labels_1, new_labels_2).any()))
             # input()
             matrix[i, j] = int(np.in1d(new_labels_1, new_labels_2).any())
-    # print('matrix ', matrix)
+    # print('matrix ', matrix[:20, :20])
+    # print('mpercentage of ones in matrix ',torch.sum(matrix)/(matrix.shape[0] * matrix.shape[1]))
+    #input()
     return matrix
